requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.9
charset-normalizer==3.4.2
colorama==0.4.6
datasets==4.0.0
dill==0.3.8
distro==1.9.0
filelock==3.13.1
frozenlist==1.7.0
fsspec==2024.6.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.4
jiter==0.10.0
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
networkx==3.3
numpy==2.1.2
openai==1.95.1
packaging==25.0
pandas==2.3.1
peft==0.16.0
pillow==11.0.0
propcache==0.3.2
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
setuptools==70.2.0
six==1.17.0
sniffio==1.3.1
sympy==1.13.3
tokenizers==0.21.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
transformers==4.53.2
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
